added nfts
